outsource mesh partitioning
